Fix review comments, function names to follow better semantics. Add distinctByKey for PairSCollectionFunctions
